加入auth中间件

+++ auth/dao/mysql/models/user.go
@@ -4,6 +4,6 @@ type UserAccount struct {
 	Id       int64  `gorm:"primaryKey;autoIncrement;not null"`
 	UserName string `gorm:"uniqueIndex;not null;type:varchar(255)"`
 	Password string `gorm:"not null;type:varchar(255)"`
-	AddTime  int
-	Status   int
+	AddTime  int64
+	Status   int8
 }
